fix[channels] moved to using Protocol instead of subclasses
